Extract composer model popover parts
