+++ src/app/components/workspace/composer/composer-model-popover.tsx
@@ -1,18 +1,17 @@
-import { Check, Search } from 'lucide-react'
-import {
-  type CSSProperties,
-  type RefObject,
-  useEffect,
-  useLayoutEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react'
+import { Search } from 'lucide-react'
+import { type CSSProperties, type RefObject, useEffect, useMemo, useRef, useState } from 'react'
 import { createPortal } from 'react-dom'
 import type { ComposerModel, ComposerThinkingLevel } from '../../../desktop/types'
-import { menuOptionClass, popoverPanelClass, toolbarButtonClass } from '../../../ui/classes'
+import { popoverPanelClass } from '../../../ui/classes'
 import { cn } from '../../../utils/cn'
 import { SurfacePanel } from '../../common/surface-panel'
+import {
+  type ComposerModelMenuOption,
+  ModelPopoverMenuList,
+  ModelPopoverTriggerButton,
+  type NestedModelMenu,
+} from './composer-model-popover-parts'
+import { useComposerModelPopoverPlacement } from './useComposerModelPopoverPlacement'
 
 type ComposerModelPopoverProps = {
   anchorRef: RefObject<HTMLButtonElement | null>
@@ -25,88 +24,6 @@ type ComposerModelPopoverProps = {
   thinkingLevelLabels: Record<ComposerThinkingLevel, string>
   onSelectModel: (model: ComposerModel) => void
   onSelectThinkingLevel: (level: ComposerThinkingLevel) => void
-}
-
-const sidePlacementGap = 8
-const sidePlacementViewportPadding = 12
-
-type NestedMenu = 'provider' | 'model' | 'thinking' | null
-
-type MenuOption = {
-  id: string
-  label: string
-  description?: string
-  selected: boolean
-  onSelect: () => void
-}
-
-function TriggerButton({
-  label,
-  value,
-  active,
-  onClick,
-}: {
-  label: string
-  value: string
-  active: boolean
-  onClick: () => void
-}) {
-  return (
-    <button
-      type="button"
-      className={cn(
-        toolbarButtonClass,
-        'grid w-full gap-0.5 rounded-xl px-2.5 py-2 text-left hover:bg-[color:var(--surface-hover)]',
-        active && 'bg-[color:var(--accent-bg-subtle)] text-[color:var(--text)]',
-      )}
-      onClick={onClick}
-    >
-      <span className="text-[11px] text-[color:var(--muted)]">{label}</span>
-      <span className="min-w-0 truncate text-[12px] text-[color:var(--text)]">{value}</span>
-    </button>
-  )
-}
-
-function MenuList({ items }: { items: MenuOption[] }) {
-  return (
-    <div
-      role="menu"
-      className={cn('-mx-1.5 -mt-1.5 pr-0', items.length > 10 && 'max-h-72 overflow-y-auto')}
-    >
-      <div className="grid min-w-0 pl-1 pt-1.5 pr-0 pb-2.5">
-        {items.map((item) => (
-          <button
-            key={item.id}
-            type="button"
-            role="menuitemradio"
-            aria-checked={item.selected}
-            className={cn(
-              menuOptionClass,
-              'mr-1 text-[12px] text-[color:var(--text)]',
-              item.selected && 'bg-[color:var(--accent-bg-subtle)]',
-            )}
-            onClick={item.onSelect}
-          >
-            <span className="inline-flex items-center justify-center text-[color:var(--accent)]">
-              {item.selected ? <Check size={13} /> : null}
-            </span>
-            <span className="min-w-0">
-              <span className="block truncate">{item.label}</span>
-              {item.description ? (
-                <span className="block truncate text-[10.5px] text-[color:var(--muted)]">
-                  {item.description}
-                </span>
-              ) : null}
-            </span>
-          </button>
-        ))}
-      </div>
-    </div>
-  )
-}
-
-function clamp(value: number, min: number, max: number) {
-  return Math.min(Math.max(value, min), max)
 }
 
 export function ComposerModelPopover({
@@ -134,39 +51,10 @@ export function ComposerModelPopover({
     })
   }, [availableModels])
 
-  const [openMenu, setOpenMenu] = useState<NestedMenu>(null)
+  const [openMenu, setOpenMenu] = useState<NestedModelMenu>(null)
   const [selectedProvider, setSelectedProvider] = useState(currentModel?.provider ?? '')
   const [modelSearch, setModelSearch] = useState('')
-  const [sidePlacementEnabled, setSidePlacementEnabled] = useState(false)
-  const [sidePosition, setSidePosition] = useState({ left: 0, top: 0 })
-  const [sidePositionReady, setSidePositionReady] = useState(false)
   const modelSearchRef = useRef<HTMLInputElement>(null)
-
-  useLayoutEffect(() => {
-    const updatePlacementMode = () => {
-      const anchorRect = anchorRef.current?.getBoundingClientRect()
-      const estimatedPanelHeight = Math.min(
-        360,
-        window.innerHeight - sidePlacementViewportPadding * 2,
-      )
-      setSidePlacementEnabled(
-        Boolean(
-          anchorRect &&
-            (preferSidePlacement ||
-              anchorRect.top <
-                estimatedPanelHeight + sidePlacementGap + sidePlacementViewportPadding),
-        ),
-      )
-    }
-
-    updatePlacementMode()
-    window.addEventListener('resize', updatePlacementMode)
-    window.addEventListener('scroll', updatePlacementMode, true)
-    return () => {
-      window.removeEventListener('resize', updatePlacementMode)
-      window.removeEventListener('scroll', updatePlacementMode, true)
-    }
-  }, [anchorRef, preferSidePlacement])
 
   useEffect(() => {
     if (currentModel?.provider) {
@@ -196,7 +84,7 @@ export function ComposerModelPopover({
   const currentModelForSelectedProvider =
     currentModel?.provider === selectedProvider ? currentModel : null
 
-  const openMenuItems = useMemo<MenuOption[]>(() => {
+  const openMenuItems = useMemo<ComposerModelMenuOption[]>(() => {
     if (openMenu === 'provider') {
       return providers.map((provider) => ({
         id: provider.provider,
@@ -257,59 +145,12 @@ export function ComposerModelPopover({
     }
   }, [showModelSearch])
 
-  useLayoutEffect(() => {
-    if (!sidePlacementEnabled) {
-      setSidePositionReady(false)
-      return
-    }
-
-    const updatePosition = (event?: Event) => {
-      const target = event?.target instanceof Node ? event.target : null
-      if (target && panelRef.current?.contains(target)) {
-        return
-      }
-
-      const anchorRect = anchorRef.current?.getBoundingClientRect()
-      const panelRect = panelRef.current?.getBoundingClientRect()
-
-      if (!(anchorRect && panelRect)) {
-        return
-      }
-
-      const maxLeft = window.innerWidth - panelRect.width - sidePlacementViewportPadding
-      const preferredLeft = anchorRect.right + sidePlacementGap
-      const left = clamp(
-        preferredLeft,
-        sidePlacementViewportPadding,
-        Math.max(sidePlacementViewportPadding, maxLeft),
-      )
-      const maxTop = window.innerHeight - panelRect.height - sidePlacementViewportPadding
-      const centeredTop = anchorRect.top + anchorRect.height / 2 - panelRect.height / 2
-      const top = clamp(
-        centeredTop,
-        sidePlacementViewportPadding,
-        Math.max(sidePlacementViewportPadding, maxTop),
-      )
-
-      setSidePosition((current) => {
-        if (current.left === left && current.top === top) {
-          return current
-        }
-
-        return { left, top }
-      })
-      setSidePositionReady(true)
-    }
-
-    updatePosition()
-    window.addEventListener('resize', updatePosition)
-    window.addEventListener('scroll', updatePosition, true)
-
-    return () => {
-      window.removeEventListener('resize', updatePosition)
-      window.removeEventListener('scroll', updatePosition, true)
-    }
-  })
+  const { sidePlacementEnabled, sidePosition, sidePositionReady } =
+    useComposerModelPopoverPlacement({
+      anchorRef,
+      panelRef,
+      preferSidePlacement,
+    })
 
   const panelContents = (
     <>
@@ -331,7 +172,7 @@ export function ComposerModelPopover({
       ) : null}
       {openMenuItems.length > 0 ? (
         <>
-          <MenuList items={openMenuItems} />
+          <ModelPopoverMenuList items={openMenuItems} />
           <div className="mx-2 mb-1 h-px bg-[color:var(--border)]" />
         </>
       ) : showModelSearch ? (
@@ -339,7 +180,7 @@ export function ComposerModelPopover({
       ) : null}
 
       <div className="relative min-w-0">
-        <TriggerButton
+        <ModelPopoverTriggerButton
           label="Provider"
           value={selectedProvider || 'Choose provider'}
           active={openMenu === 'provider'}
@@ -351,7 +192,7 @@ export function ComposerModelPopover({
       </div>
 
       <div className="relative min-w-0">
-        <TriggerButton
+        <ModelPopoverTriggerButton
           label="Model"
           value={
             currentModelForSelectedProvider?.name ?? modelsForProvider[0]?.name ?? 'Choose model'
@@ -371,7 +212,7 @@ export function ComposerModelPopover({
       </div>
 
       <div className="relative min-w-0">
-        <TriggerButton
+        <ModelPopoverTriggerButton
           label="Reasoning"
           value={thinkingLevelLabels[currentThinkingLevel]}
           active={openMenu === 'thinking'}
